backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
alembic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-jose
bcrypt==4.0.1
